🔧 Set cors header in next.config
