+++ next.config.mjs
@@ -13,6 +13,28 @@ const nextConfig = {
             'cdn.pixabay.com',
         ],
     },
+
+    async headers() {
+        return [
+            {
+                source: '/api/:path*',
+                headers: [
+                    {
+                        key: 'Access-Control-Allow-Origin',
+                        value: '*',
+                    },
+                    {
+                        key: 'Access-Control-Allow-Methods',
+                        value: 'GET, POST, PUT, DELETE, OPTIONS',
+                    },
+                    {
+                        key: 'Access-Control-Allow-Headers',
+                        value: 'Content-Type, Authorization',
+                    },
+                ],
+            },
+        ]
+    },
 }
 
 export default nextConfig
